extensions: add support for decoding RSA public keys
The :rsa-public-key type can be used to decode an RSA public key from
a given binary stream, where in the stream the data that follows is
the actual public key components. Generally this would be used when
decoding the public key of an OpenSSH certificate, which embeds an RSA
public key, since the public key is not preceeded by a `string` value,
which specifies the key kind.

The :ssh-rsa-public-key type can be used to decode an RSA public key,
which is encoded as per RFC 4253, which first encodes the key
kind (e.g. the "ssh-rsa" string), and then the public key parts - the
exponent and the modulus.