Remove `extensions` system
The `extensions` system is probably going to evolve on it's own over
time, e.g. adding more public/private key pairs support, certificates,
signing of public keys, etc. and most of these things are covered as
part of various other RFC documents, so it is better to have these
bits taken out of the `cl-rfc4251` system as a whole, just to avoid
potential confusion for users.

As of now the `cl-rfc4251` system supports all data types defined in
RFC 4251 -- both encoding and decoding, so we can consider this one
feature complete.